# assignment6_demo/artifacts/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ISA-D Demo Backend",
    "description": "Tìm sản phẩm bằng tiếng Việt, ảnh, text + ảnh; Azure Speech và tra đơn hàng demo.",
    "version": "0.1.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/meta": {
      "get": {
        "summary": "Meta",
        "operationId": "meta_api_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "summary": "Products",
        "operationId": "products_api_v1_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "summary": "Product Detail",
        "operationId": "product_detail_api_v1_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits": {
      "get": {
        "summary": "Credits",
        "operationId": "credits_api_v1_credits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/media/products/{product_id}": {
      "get": {
        "summary": "Media",
        "operationId": "media_api_v1_media_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "summary": "Order Query",
        "operationId": "order_query_api_v1_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "summary": "Order Path",
        "operationId": "order_path_api_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "summary": "Text Search",
        "operationId": "text_search_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/image": {
      "post": {
        "summary": "Image Search",
        "operationId": "image_search_api_v1_search_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "options": {
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "image"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/search/multimodal": {
      "post": {
        "summary": "Multimodal Search",
        "operationId": "multimodal_search_api_v1_search_multimodal_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "text": {
                    "type": "string"
                  },
                  "text_weight": {
                    "type": "string"
                  },
                  "options": {
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "image"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/speech/transcriptions": {
      "post": {
        "summary": "Transcribe",
        "operationId": "transcribe_api_v1_speech_transcriptions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary"
                  },
                  "language": {
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "audio"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Filters": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "running_shoes",
                  "trail_shoes",
                  "casual_shoes",
                  "bag"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "min_price": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price"
          },
          "max_price": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price"
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Filters"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SearchOptions": {
        "properties": {
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "result_policy": {
            "type": "string",
            "enum": [
              "nearest",
              "relevant"
            ],
            "title": "Result Policy",
            "default": "nearest"
          },
          "filters": {
            "$ref": "#/components/schemas/Filters"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SearchOptions"
      },
      "TextRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "text",
              "voice"
            ],
            "title": "Mode"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "voice_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "azure",
                  "manual_transcript"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Source"
          },
          "options": {
            "$ref": "#/components/schemas/SearchOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "text"
        ],
        "title": "TextRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}